requirements.txt
altgraph==0.17.4
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Django==5.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-rest-framework==0.1.0
django-tqdm==1.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
flake8-html==0.4.3
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
Markdown==3.5.1
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.0
pefile==2023.2.7
pillow==10.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
PyJWT==2.9.0
pyproject-flake8==7.0.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
setuptools==69.5.1
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2024.1
Werkzeug==3.0.4
